Remove the dependency on Guava for easier j2cl compatibility.
At this time, guava-gwt doesn't work in j2cl as it is in maven, since
you need to include both guava and guava-gwt on your classpath to compile
correctly in GWT2. Changing this could be done in a few ways, like providing
two different gwt-safecss poms (once a guava-j2cl existed), or since we only
need guava for a one method and one annotation, we can simply cut the
dependency outright.